Split ConfigFragments.scala into multiple files by usage
