+++ docs/Generators/SiFive-Generators.rst
@@ -18,7 +18,7 @@ Peripheral Devices
 These peripheral devices usually affect the memory map of the SoC, and its top-level IO as well.
 To integrate one of these devices in your SoC, you will need to define a custom config fragment with the approriate address for the device using the Rocket Chip parameter system. As an example, for a GPIO device you could add the following config fragment to set the GPIO address to ``0x10012000``. This address is the start address for the GPIO configuration registers.
 
-.. literalinclude:: ../../generators/chipyard/src/main/scala/ConfigFragments.scala
+.. literalinclude:: ../../generators/chipyard/src/main/scala/config/fragments/PeripheralFragments.scala
     :language: scala
     :start-after: DOC include start: gpio config fragment
     :end-before: DOC include end: gpio config fragment
